+++ packages/scenes-app/src/pages/Demos/getDemos.ts
@@ -1,5 +1,10 @@
-import { EmbeddedScene, SceneCanvasText, SceneFlexLayout } from '@grafana/scenes';
+import { EmbeddedScene, SceneCanvasText, SceneFlexItem, SceneFlexLayout } from '@grafana/scenes';
+import { getFlexLayoutTest } from './scenes/flexLayout';
+import { getGridLayoutTest } from './scenes/grid';
+import { getGridWithRowLayoutTest } from './scenes/gridWithRow';
+import { getNestedScene } from './scenes/nestedScene';
 import { getPanelMenuTest } from './scenes/panelMenu';
+import { getPanelRepeaterTest } from './scenes/panelRepeater';
 import { getQueryEditorDemo } from './scenes/queryEditor';
 
 interface SceneDef {
@@ -10,6 +15,11 @@ interface SceneDef {
 export function getDemos(): SceneDef[] {
   return [
     { title: 'Panel menu', getScene: getPanelMenuTest },
+    { title: 'Flex layout', getScene: getFlexLayoutTest },
+    { title: 'Panel repeater', getScene: getPanelRepeaterTest },
+    { title: 'Grid layout', getScene: getGridLayoutTest },
+    { title: 'Grid layout with rows', getScene: getGridWithRowLayoutTest },
+    { title: 'Nested scene', getScene: getNestedScene },
     { title: 'Query editor demo', getScene: getQueryEditorDemo },
   ];
 }
@@ -19,9 +29,11 @@ function getErrorScene(title: string) {
     body: new SceneFlexLayout({
       direction: 'column',
       children: [
-        new SceneCanvasText({
-          text: `¯\\_(ツ)_/¯ No scene found with title ${title}`,
-          align: 'center',
+        new SceneFlexItem({
+          body: new SceneCanvasText({
+            text: `¯\\_(ツ)_/¯ No scene found with title ${title}`,
+            align: 'center',
+          }),
         }),
       ],
     }),