+++ packages/scenes/src/components/NestedScene.tsx
@@ -6,7 +6,13 @@ import { Stack } from '@grafana/experimental';
 import { Button, ToolbarButton, useStyles2 } from '@grafana/ui';
 
 import { SceneObjectBase } from '../core/SceneObjectBase';
-import { SceneObject, SceneLayoutChildState, SceneComponentProps, SceneLayout } from '../core/types';
+import {
+  SceneObject,
+  SceneLayoutChildState,
+  SceneComponentProps,
+  SceneLayout,
+  SceneLayoutItemState,
+} from '../core/types';
 
 interface NestedSceneState extends SceneLayoutChildState {
   title: string;
@@ -40,6 +46,12 @@ export class NestedScene extends SceneObjectBase<NestedSceneState> {
     if ('children' in parent.state) {
       parent.setState({
         children: parent.state.children.filter((x) => x !== this),
+      });
+    }
+
+    if (isSceneLayoutItem(parent)) {
+      parent.setState({
+        body: undefined,
       });
     }
   };
@@ -115,3 +127,7 @@ const getStyles = (theme: GrafanaTheme2) => ({
     flexGrow: 1,
   }),
 });
+
+function isSceneLayoutItem(x: SceneObject): x is SceneObject<SceneLayoutItemState> {
+  return 'body' in x.state;
+}